move usePending
