+++ apps/client/src/app/account/trusted-users/_component/TrustedWrapper.tsx
@@ -6,8 +6,8 @@ import TrustedList from './TrustedList';
 import TrustedByList from './TrustedByList';
 import { UseQueryResult } from '@tanstack/react-query';
 import { ResultOf } from '@graphql-typed-document-node/core';
-import { useCallback, useMemo, useState } from 'react';
-import { IsPending, RunPending, usePending } from '../_util/usePending';
+import { useCallback, useMemo } from 'react';
+import { IsPending, RunPending, usePending } from '@/util/usePending';
 import { prepMemberList } from '../_util/prepMemberList';
 import { useUser } from '@/app/_ctx/user/context';
 import { prettyErrorPlaceholder } from '@/util/prettyErrors';
